Promote NoOpRetainedStateRegistry to public API
